Fix bizarre unit test breakage
git-svn-id: http://svn.dmdirc.com/trunk@3805 00569f92-eb28-0410-84fd-f71c24880f43

+++ test/com/dmdirc/config/ConfigManagerTest.java
@@ -37,7 +37,7 @@ public class ConfigManagerTest {
         assertNull(ConfigManager.getStats().get("unit-test123.baz"));
         cm.hasOption("unit-test123", "baz");
         assertNotNull(ConfigManager.getStats().get("unit-test123.baz"));
-        assertEquals(1, ConfigManager.getStats().get("unit-test123.baz"));
+        assertEquals(1, (int) ConfigManager.getStats().get("unit-test123.baz"));
     }
     
     public static junit.framework.Test suite() {
